new parameter enable or disable recursive managing of env dir
When setting recursive to true for managing all the files under
the environment dir, puppet runs can be very long if there are
a lot of environments available. This setting can turn off the
recursive managing of the environments dir.
The first idea was to chose if the owner and group should be managed,
but that does not make sense. It's better to have the choice of
managing the recursive flag.